Tons 'o Changes.
The menu, tree, and selectGrid have been moved into FTPBrowser, but several things need to be fixed up still regarding the whole inner working of that stuff (it would help if I planned this stuff out ahead of time).

FTPFileItem is now being fully utilized by the FTPTreeItem (PHP should send back full paths, need to comment this elsewhere).

A couple UI things now feel snappier because of the use of DeferredCommand, yay!

Still lots of cleanup to be done for some reason